+++ portal/src/screen/AuthenticationForm/Signup/style.js
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+import { ButtonRoot } from "../style";
+
+export const ButtonGoogle = styled(ButtonRoot)`
+  background-color: #e5e5e5;
+  color: black;
+`;
+
+export const ButtonFacebook = styled(ButtonRoot)`
+  background-color: #3b5998;
+`;
+
+export const ButtonWrapper = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+`;
